new mut.instruments
Previously, the mut.instruments contained a couple of prototypes
for Unit-Generator-like functions, similar to pink.instruments.

These things were relatively low-level, and used mutable data
to implement things. Not very clojury.

So, this commit drops the old implementation, and replaces it
with a new higher-level concept.

Now there is the MusicObject->AudioEngineEvnets protocol,
and instrument modules are supposed to implement the protocol,
and thus the instruments now are about converting abstract
music objects to audio functions.

So the mut.instruments is kinda separate application layer now.
Not very much code in there, just abstract definitions for future.