re-organize modules: move code from audio/engine.clj -> audio.clj
No changes in logic, just move the code.

Because:

1. The audio.clj is a good starting point for reading the code,
and thus it is good to put there an `Orchestra` - the main global
list that refers to all other instruments and utils in sub-modules.

2. I don't quite like how code is currently organized.
So I'm going to start over again by putting the majority of
existing code to audio.clj, and then evicting "utils" to sub-modules
as they are written.