get rid of Orchestra "object" in favor of global vars
- the `mut.audio/Orchestra` record definition is dropped
- many functions in the `mut.audio` module lost their 1st `orchestra`
  parameter, and instead now they just refer to several global vars

Some time ago I invented this "Orchestra" abstraction.
It represented the state: which instruments are currently allocated,
and preferences for allocation new instruments.

I though it is a good idea to pass the state as an argument to all
functions in the `audio` module, but in reality, it brought more pain
than benefit, especially in REPL. It turns out that I always end up
with having some global or almost global singleton variable,
that I have to pass as an argument everywhere.

So eventually I decided to get rid of these `Orchestra` singleton.
All functions in the `mut.audio` module lost the 1st `orchestra`
parameter. It became worse for unit tests, but now I can use the
functions immediately after I start REPL.

This is valuable not only for REPL, but also for any kind of
protoyping the code. Now I don't have to initialize the complex state
before I can actually produce the sound.